+++ src/lib.ts
@@ -32,6 +32,7 @@ import { validateDuplicateContextOrInfo } from "./validations/validateDuplicateC
 import { validateSemanticNullability } from "./validations/validateSemanticNullability";
 import { resolveTypes } from "./transforms/resolveTypes";
 import { resolveResolverParams } from "./transforms/resolveResolverParams";
+import { propagateHeritage } from "./transforms/propagateHeritage";
 
 // Export the TypeScript plugin implementation used by
 // grats-ts-plugin
@@ -115,6 +116,8 @@ export function extractSchemaAndDoc(
         .andThen((doc) => applyDefaultNullability(doc, config))
         // Merge any `extend` definitions into their base definitions.
         .map((doc) => mergeExtensions(doc))
+        // Add fields from extended classes and implemented interfaces.
+        .map((doc) => propagateHeritage(ctx, doc))
         // Sort the definitions in the document to ensure a stable output.
         .map((doc) => sortSchemaAst(doc))
         .result();