+++ src/tests/fixtures/inheritance/classInheritsFieldFromAbstractParent.ts
@@ -1,0 +1,11 @@
+/** @gqlType */
+abstract class Parent {
+  /** @gqlField */
+  parentField: string;
+}
+
+/** @gqlType */
+export class Child extends Parent {
+  /** @gqlField */
+  childField: string;
+}