+++ src/tests/fixtures/interfaces/interfaceNamedQuery.invalid.ts
@@ -1,0 +1,5 @@
+/** @gqlInterface */
+interface Query {
+  /** @gqlField */
+  field: string;
+}