+++ website/docs/04-docblock-tags/snippets/04-interface-class-declaration.grats.ts
@@ -1,0 +1,7 @@
+/** @gqlInterface */
+class MyInterfaceClass {
+  /** @gqlField */
+  someField(): string {
+    return "someField";
+  }
+}